+++ chat2db-client/.umirc.prod.ts
@@ -1,8 +1,8 @@
 import { defineConfig } from 'umi';
 import { extractYarnConfig } from './src/utils/webpack';
 const MonacoWebpackPlugin = require('monaco-editor-webpack-plugin');
-const UMI_PublicPath = process.env.UMI_PublicPath || './static/front/';
 const yarn_config = extractYarnConfig(process.argv);
+const publicPath = yarn_config.public_path || './static/front/';
 
 const chainWebpack = (config: any, { webpack }: any) => {
   config.plugin('monaco-editor').use(MonacoWebpackPlugin, [
@@ -13,7 +13,7 @@ const chainWebpack = (config: any, { webpack }: any) => {
 };
 
 export default defineConfig({
-  publicPath: UMI_PublicPath,
+  publicPath: publicPath,
   chainWebpack,
   define: {
     'process.env.UMI_ENV': process.env.UMI_ENV,