+++ chat2db-client/src/components/Console/index.tsx
@@ -176,42 +176,47 @@ function Console(props: IProps) {
 
   const handleApiKeyEmptyOrGetQrCode = async (shouldPoll?: boolean) => {
     setIsLoading(true);
-    const { wechatQrCodeUrl, token, tip } = await aiServer.getLoginQrCode({});
-    setIsLoading(false);
+    try {
+      const { wechatQrCodeUrl, token, tip } = await aiServer.getLoginQrCode({});
+      setIsLoading(false);
 
-    setPopularizeModal(true);
-    setModalProps({
-      imageUrl: wechatQrCodeUrl,
-      token,
-      tip,
-    });
-    if (shouldPoll) {
-      let pollCnt = 0;
-      aiFetchIntervalRef.current = setInterval(async () => {
-        const { apiKey } = (await aiServer.getLoginStatus({ token })) || {};
-        pollCnt++;
-        if (apiKey || pollCnt >= 60) {
-          clearInterval(aiFetchIntervalRef.current);
-        }
-        if (apiKey) {
-          setPopularizeModal(false);
+      setPopularizeModal(true);
+      setModalProps({
+        imageUrl: wechatQrCodeUrl,
+        token,
+        tip,
+      });
+      if (shouldPoll) {
+        let pollCnt = 0;
+        aiFetchIntervalRef.current = setInterval(async () => {
+          const { apiKey } = (await aiServer.getLoginStatus({ token })) || {};
+          pollCnt++;
+          if (apiKey || pollCnt >= 60) {
+            clearInterval(aiFetchIntervalRef.current);
+          }
+          if (apiKey) {
+            setPopularizeModal(false);
 
-          await dispatch({
-            type: 'ai/setAiConfig',
-            payload: {
-              ...(aiModel.aiConfig || {}),
-              apiKey,
-            },
-          });
-          await dispatch({
-            type: 'ai/fetchRemainingUse',
-            payload: {
-              apiKey,
-            },
-          });
-        }
-      }, 3000);
+            await dispatch({
+              type: 'ai/setAiConfig',
+              payload: {
+                ...(aiModel.aiConfig || {}),
+                apiKey,
+              },
+            });
+            await dispatch({
+              type: 'ai/fetchRemainingUse',
+              payload: {
+                apiKey,
+              },
+            });
+          }
+        }, 3000);
+      }
+    }catch (e) {
+      setIsLoading(false);
     }
+
   };
 
   const handleAIChatInEditor = async (content: string, promptType: IPromptType) => {
@@ -437,7 +442,7 @@ function Console(props: IProps) {
           onExecute={executeSQL}
           options={props.editorOptions}
           tables={props.tables}
-          // onChange={}
+        // onChange={}
         />
         {/* <Modal open={modelConfig.open}>{modelConfig.content}</Modal> */}
         <Drawer open={isAiDrawerOpen} getContainer={false} mask={false} onClose={() => setIsAiDrawerOpen(false)}>
@@ -466,8 +471,17 @@ function Console(props: IProps) {
         <Button
           type="text"
           onClick={() => {
-            const contextTmp = editorRef?.current?.getAllContent();
-            editorRef?.current?.setValue(format(contextTmp || ''), 'cover');
+            let formatRes = '';
+            try {
+              const contextTmp = editorRef?.current?.getAllContent();
+              formatRes = format(contextTmp || '');
+            }
+            catch { }
+            if (formatRes) {
+              editorRef?.current?.setValue(formatRes, 'cover');
+            } else {
+              message.error(i18n('common.tips.formatError'))
+            }
           }}
         >
           {i18n('common.button.format')}