+++ chat2db-server/chat2db-server-domain/chat2db-server-domain-api/src/main/java/ai/chat2db/server/domain/api/service/ViewService.java
@@ -1,5 +1,6 @@
 package ai.chat2db.server.domain.api.service;
 
+import ai.chat2db.server.tools.base.wrapper.result.DataResult;
 import ai.chat2db.server.tools.base.wrapper.result.ListResult;
 import ai.chat2db.spi.model.Table;
 import jakarta.validation.constraints.NotEmpty;
@@ -19,4 +20,11 @@ public interface ViewService {
     ListResult<Table> views(@NotEmpty String databaseName, String schemaName);
 
 
+    /**
+     * Querying the details of a view.
+     *
+     * @param databaseName
+     * @return
+     */
+    DataResult<Table> detail(@NotEmpty String databaseName, String schemaName,String tableName);
 }