+++ chat2db-server/chat2db-server-web/chat2db-server-web-api/src/main/java/ai/chat2db/server/web/api/controller/operation/saved/request/OperationUpdateRequest.java
@@ -38,4 +38,9 @@ public class OperationUpdateRequest {
      * 是否在tab中被打开,y表示打开,n表示未打开
      */
     private String tabOpened;
+
+    /**
+     * operation type
+     */
+    private String operationType;
 }