+++ chat2db-server/chat2db-server-web/chat2db-server-web-api/src/main/java/ai/chat2db/server/web/api/controller/rdb/request/FunctionDetailRequest.java
@@ -1,0 +1,34 @@
+package ai.chat2db.server.web.api.controller.rdb.request;
+
+import ai.chat2db.server.web.api.controller.data.source.request.DataSourceBaseRequestInfo;
+import jakarta.validation.constraints.NotNull;
+import lombok.Data;
+
+@Data
+public class FunctionDetailRequest implements DataSourceBaseRequestInfo {
+
+    /**
+     * 数据源id
+     */
+    @NotNull
+    private Long dataSourceId;
+    /**
+     * DB名称
+     */
+    private String databaseName;
+
+    /**
+     * 表所在空间，pg,oracle需要，mysql不需要
+     */
+    private String schemaName;
+
+    /**
+     * function name
+     */
+    private String functionName;
+
+    /**
+     * if true, refresh the cache
+     */
+    private boolean refresh;
+}