+++ chat2db-server/chat2db-spi/src/main/java/ai/chat2db/spi/MetaData.java
@@ -15,8 +15,8 @@ import jakarta.validation.constraints.NotEmpty;
 public interface MetaData {
     /**
      * Query all databases.
-     * @param connection
      *
+     * @param connection
      * @return
      */
     List<Database> databases(Connection connection);
@@ -51,6 +51,18 @@ public interface MetaData {
      * @return
      */
     List<Table> tables(Connection connection, @NotEmpty String databaseName, String schemaName, String tableName);
+
+    /**
+     * Querying view information.
+     *
+     * @param connection
+     * @param databaseName
+     * @param schemaName
+     * @param viewName
+     * @return
+     */
+    Table view(Connection connection, @NotEmpty String databaseName, String schemaName, String viewName);
+
 
     /**
      * Querying all views under a schema.
@@ -128,4 +140,36 @@ public interface MetaData {
     List<TableIndex> indexes(Connection connection, @NotEmpty String databaseName, String schemaName,
         @NotEmpty String tableName);
 
+    /**
+     * Querying function detail under a schema.
+     *
+     * @param connection
+     * @param databaseName
+     * @param schemaName
+     * @param functionName
+     * @return
+     */
+    Function function(Connection connection, @NotEmpty String databaseName, String schemaName, String functionName);
+
+    /**
+     * Querying  trigger  under a schema.
+     *
+     * @param connection
+     * @param databaseName
+     * @param schemaName
+     * @param triggerName
+     * @return
+     */
+    Trigger trigger(Connection connection, @NotEmpty String databaseName, String schemaName, String triggerName);
+
+    /**
+     * Querying all procedures under a schema.
+     *
+     * @param connection
+     * @param schemaName
+     * @param databaseName
+     * @param procedureName
+     * @return
+     */
+    Procedure procedure(Connection connection, @NotEmpty String databaseName, String schemaName,String procedureName);
 }